Fix build of gst-plugins
Minor changes to deploy script
Fixes to launch script to handle TZ and start ntpd if not already
started

+++ scripts/build-docker.sh
@@ -3,6 +3,7 @@
 : ${SH:=0}
 : ${STATIC:=0}
 : ${PROF:=0}
+: ${BLOB_LIGHT:=0}
 
 : ${ARCH:="amd64"}
 if [ -n "${1}" ]; then
@@ -37,8 +38,13 @@ if [ ${SH} -eq 1 ]; then
 fi
 
 # copy the artifacts to the output directory
-if [ ${BLOB} -eq 1 ]; then
+if [ ${BLOB} -eq 1 ] && [ ${BLOB_LIGHT} -eq 0 ]; then
 	docker run --rm -v $(pwd)/build:/out --security-opt label=disable -e ARCH=${ARCH} genie-builder:${ARCH} /src/scripts/blob.sh
+	exit ${?}
+fi
+
+if [ ${BLOB_LIGHT} -eq 1 ]; then
+	docker run --rm -v $(pwd)/build:/out --security-opt label=disable -e ARCH=${ARCH} genie-builder:${ARCH} /src/scripts/binonly.sh
 	exit ${?}
 fi
